Install the findmodules too
These weren't installed by default, probably since the use of
sane_install (the irony!).

They are now installed to sane_install's cmake_module_path/modules, but
this value is hardcoded. It would be nice to get them from sane_install
itself, but I wasn't sure how to do it. The interface should be a
key-value map (library -> "lib/foo", binary -> "bin/"), but how to get
this value is a bit difficult:

- set all the paths in the PARENT_SCOPE (hide them from the cache). The
  user can then get them with ${sane_install_library_path} etc.

  This is a bit annoying: can't use it until after sane_install has been
  called, and a bit error prone, since multiple calls to sane_install
  will overwrite the values.

- Create a special funtion that calculates all the paths it
  would use (would need the same arguments as the sane_install command)
  and sets them in a parent_scope or fills a given variable/set of
  variables:

  get_sane_install_paths( ... args to sane_install ... GET_LIB_PATH
  libpath) -> libpath = "lib"